adding user pay model

+++ Ascendix-Backend/Dto/UserCourseProgressDto/UpdateUserCourseProgress.cs
@@ -9,7 +9,7 @@ namespace Ascendix_Backend.Dto.UserCourseProgressDto
     public class UpdateUserCourseProgress
     {
         public Guid? courseId { get; set; }
-        public decimal? progressPercent { get; set; }
+        public int? progressPercent { get; set; }
         public Status? status { get; set; }
 
     }
