+++ Ascendix-Backend/Dto/UserCourseProgressDto/ViewUserCourseProgress.cs
@@ -11,7 +11,7 @@ namespace Ascendix_Backend.Dto.UserCourseProgressDto
         public Guid progressId { get; set; }
         public string? userId { get; set; }
         public Guid courseId { get; set; }
-        public decimal progressPercent { get; set; }
+        public int progressPercent { get; set; }
         public Status status { get; set; }
         
     }
