+++ Ascendix-Backend/Dto/UserModuleDto/ViewUserModule.cs
@@ -12,6 +12,6 @@ namespace Ascendix_Backend.Dto.UserModuleDto
         public Guid moduleId { get; set; }
         public string? userId { get; set; }
         public Status status { get; set; }
-        public decimal progressPercent { get; set; }
+        public int progressPercent { get; set; }
     }
 }
