+++ Ascendix-Backend/Models/UserModule.cs
@@ -11,7 +11,7 @@ namespace Ascendix_Backend.Models
         public Guid moduleId { get; set; }
         public string? userId { get; set; }
         public Status status { get; set; }
-        public decimal progressPercent { get; set; }
+        public int progressPercent { get; set; }
 
         public Module? modules { get; set; }
         public User? user { get; set; }
